+++ src/query/storages/fuse/src/operations/mutation/mod.rs
@@ -15,14 +15,14 @@
 pub mod abort_operation;
 pub mod base_mutator;
 mod compact;
-pub mod mutation_fill_internal_columns;
-pub mod mutation_meta;
+mod mutation_aggregator;
+mod mutation_fill_internal_columns;
+mod mutation_meta;
 mod mutation_part;
-pub mod mutation_sink;
+mod mutation_serialize_data;
+mod mutation_sink;
 mod mutation_source;
-pub mod recluster_mutator;
-mod transform_mutation_aggregator;
-mod transform_serialize_data;
+mod recluster_mutator;
 
 pub use abort_operation::AbortOperation;
 pub use base_mutator::BaseMutator;
@@ -33,15 +33,15 @@ pub use compact::CompactSource;
 pub use compact::SegmentCompactMutator;
 pub use compact::SegmentCompactionState;
 pub use compact::SegmentCompactor;
+pub use mutation_aggregator::MutationAggregator;
 pub use mutation_fill_internal_columns::FillInternalColumnProcessor;
 pub use mutation_meta::Mutation;
 pub use mutation_meta::MutationSinkMeta;
 pub use mutation_meta::MutationTransformMeta;
 pub use mutation_meta::SerializeDataMeta;
 pub use mutation_part::MutationPartInfo;
+pub use mutation_serialize_data::SerializeDataTransform;
 pub use mutation_sink::MutationSink;
 pub use mutation_source::MutationAction;
 pub use mutation_source::MutationSource;
 pub use recluster_mutator::ReclusterMutator;
-pub use transform_mutation_aggregator::MutationAggregator;
-pub use transform_serialize_data::SerializeDataTransform;
