+++ src/query/storages/fuse/src/operations/mutation/mutation_serialize_data.rs
